Repository -> JPA Repository로 변경, 기존 hashmap Repository -> ProductMapRepository로 변경
